feat: implement search bar and input components
