+++ src/common/Input/input.scss
@@ -1,0 +1,16 @@
+.customInput {
+  display: flex;
+  height: 35px;
+  border-radius: 0.25rem;
+  border: 2px solid #007ee5;
+
+  .input {
+    border: none;
+    padding-left: 0.5rem;
+    font-size: 1.5rem;
+
+    &:focus-visible {
+      outline: 0;
+    }
+  }
+}
