+++ src/components/Courses/components/CourseCard/course-card.scss
@@ -1,10 +1,10 @@
 .CourseCard {
-  outline: 2px solid blue;
-  width: 90%;
   display: flex;
   column-gap: 1.5rem;
-  margin: 1rem auto;
   padding: 1rem;
+  margin-bottom: 1.5rem;
+  box-shadow: 0 0 4px 3px rgba(0,0,0,0.32);
+  border-radius: 0.5rem;
 
   .title{
     font-size: 2rem;
