+++ src/components/Courses/components/SearchBar/search-bar.scss
@@ -1,0 +1,8 @@
+.search-bar {
+  width: 100%;
+  display: flex;
+  column-gap: 1rem;
+  align-items: center;
+  margin-bottom: 3rem;
+
+}
