+++ src/components/Courses/courses.scss
@@ -1,5 +1,11 @@
 .courses {
   flex: 1 0 auto;
-  border: 2px solid red;
-  margin: 1rem;
+  //border: 2px solid red;
+  margin: 0 1rem;
+  padding: 1rem;
+
+  .coursesControlsWrapper {
+    display: flex;
+    justify-content: space-between;
+  }
 }
