+++ src/components/Header/header.scss
@@ -4,7 +4,7 @@
   justify-content: space-between;
   margin: 1rem;
   padding: 0.5rem 1rem;
-  border: 2px solid pink;
+  box-shadow: 0 0 4px 3px rgba(0,0,0,0.32);
 
   .buttonWrapper {
     width: 20%;
